+++ config/laravel-webdav-server-filament.php
@@ -2,5 +2,7 @@
 
 // config for N3XT0R/LaravelWebdavServerFilament
 return [
-
+    'notifications' => [
+        'enabled' => true,
+    ],
 ];